+++ assets/src/components/activities/file_upload/actions.ts
@@ -1,0 +1,9 @@
+import { FileUploadSchema, FileSpec } from './schema';
+
+export const FileUploadActions = {
+  editFileSpec(fileSpec: FileSpec) {
+    return (model: FileUploadSchema) => {
+      model.fileSpec = fileSpec;
+    };
+  },
+};